update inheritance for better support for nested properties / conditional fields
